Explained `alarm()` everywhere. Updated other comments for clarity. Moved initialization code closer to the other initialization code.
Bug explanation:

Back in the misty deeps of time (far enough back that we couldn't dig it out of
revision control in a reliable way), we think that the test_s2c test used
`alarm(10)` as its ten-second timer. To do this, it needed to install its own
SIGALRM handler that was a no-op. Over time, the code was upgraded into a
`while(time() < end_time)`-style loop, but the signal handling code was never
removed.  Then, after the introduction of SSL, it became possible for test_s2c
to cause a `read()` to occur in e.g. SSL renegotiation or the like.

This left us in the situation where if the client disappeared off the network
(never sent a FIN that we received) when we were blocked in a `read()` call,
the child process would just stay hung forever.  This bug was causing real
production pain for MLab, so we finally tracked it down and we have tried to
prevent it in a "belt & suspenders" manner by both ensuring that the watchdog
timer is never disabled during test_s2c or test_c2s, and by installing socket
timeouts on all communication sockets.

As with all changes to the NDT codebase, this small amount of code required a
lot of archeological detective work.